Moved navigation creation into a hook
